Fix bugs in setAngleInRangeZeroTwoPi (#2011)
The function would compute the angle modulo 360, not 2pi as the function's name promises. Additionally, [fmod](https://en.cppreference.com/w/cpp/numeric/math/fmod) returns negative values if the argument is negative, which we also don't want. This can be fixed by adding 2pi to the argument - assuming the angle is not lower than -2pi.